category-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
motor
python-jose[cryptography]